more tests for god of tests

+++ include/test.h
@@ -4,6 +4,6 @@
 #include "functions.h"
 
 void CompressDecompressTest(const string& test_filename, const string& compressed_filename,
-                            const string& decompressed_filename, const int& expected_ratio);
+                            const string& decompressed_filename, const size_t& expected_ratio);
 
 #endif //MY_PROJECT_TEST_H
